After a lot of thought, remove the todo's around pattern refinement
I've decided this is not the correct way forward. Instead, it would be better to get copies of the new files and those files then added to a master directory to create a larger sample set. My ideal would be for the community to build a repo with sample files which can then be collectively used to build patterns.

Refinement would be difficult here since there are complexities with refining the strings - it's impossible to do a like-for-like comparison with the strings and so it makes the process somewhat problematic.